Improve signal handling util package
- Move code to a separate package, mirroring upstream, and separating
  code out of the generic util package.
- Default to os.Exit(2) instead of os.Exit(0), matching the default
  behavior of the Go runtime.
- Unify list of termination signals across supported platforms.
